default security group name is undef when it does not exist
The BMO extension provides a default security group to products. I think
this is a good feature that should be core'd. In the meantime, there is
no default security group for the initial data set, so this makes it so
that when there isn't one, when the name is looked up it isn't an error.