this changes makes api_field_* variables lazy
api_field_types and api_field_names rely on values from the
Bugzilla->fields definition. This in turn requires checking the schema
of the bugs table at run time. That means trying to access these when a
database is not configured, or before schema migrations have happened is
an error.

This changes those variables to hold functions (sub refs) that are
comptued when called. The help performance not degrade, we make use of
the Bugzilla::request_cache to only re-compute this once per request.