compose: mariadb -> service_healthy
MARIADB_ALLOW_EMPTY_PASSWORD isn't a variable recognised by
the MariaDB container, and MARIADB_ALLOW_EMPTY_ROOT_PASSWORD is
a bit false when we set a password.

Also use MariaDB's healthcheck script to ensure the service is
healthy before running tests. There is other mechanisms to
enforce this, but lets be clear at the basic level of the healthcheck.sh
script maintained in the container (currently by me).